Ajustado o caminho de confirmação de email para todos os users, verificação de duplicidade de email de revendedores, ajueste no modelo de admin com remoção de retornos invalidos, e a mudanca de retorno na listagem de clientes pelos admins

+++ src/models/clientes_models.py
@@ -76,7 +76,7 @@ class ClienteResponse(BaseModel):
     email: str
     cod_confirmacao_email: str
     criacao_de_conta: str
-    pontos_fidelidade: int
+    pontos_fidelidade: float
     clube_fidelidade: bool
     cod_indicacao: int
     status: bool
